requirements.txt
hid==1.0.5
numpy-stl==3.0.0
numpy>=1.26.4
opencv-python>=4.10.0.82
PyGLM>=2.7.1
PyOpenGL==3.1.6
PyQt5-Qt5==5.15.2
PyQt5-sip>=12.15.0
PyQt5-stubs==5.15.6.0
PyQt5==5.15.11
pyserial==3.5
pyserial-asyncio==0.6
python-utils==3.4.5
qasync==0.27.1
setuptools==67.0.0
wheel==0.38.4
hidapi>=0.14.0
